Tidy up: fixes; improve verbose-mode messages; update docs & CLI
