+++ tests/cylc-poll/17-pbs-cant-connect.t
@@ -31,7 +31,7 @@ then
     skip_all "\"[test battery][batch systems][${BATCH_SYS_NAME}]host\" not defined"
 fi
 
-set_test_number 3
+set_test_number 4
 install_suite "${TEST_NAME_BASE}" "${TEST_NAME_BASE}"
 if [[ "${CYLC_TEST_BATCH_TASK_HOST}" != 'localhost' ]]; then
     ssh -n "${CYLC_TEST_BATCH_TASK_HOST}" "mkdir -p 'cylc-run/${SUITE_NAME}/'"
@@ -41,11 +41,17 @@ fi
 run_ok "${TEST_NAME_BASE}-validate" cylc validate "${SUITE_NAME}"
 suite_run_ok "${TEST_NAME_BASE}-run" \
     cylc run --reference-test --debug --no-detach "${SUITE_NAME}"
-sed -n 's/^.*\(\[jobs-poll err\] Connection refused\).*$/\1/p;
+# ssh security warnings may appear between outputs => check separately too.
+sed -n 's/^.*\(\[jobs-poll err\]\) \(Connection refused\).*$/\1\n\2/p;
+        s/^.*\(\[jobs-poll err\]\).*$/\1/p;
+        s/^.*\(Connection refused\).*$/\1/p;
         s/^.*\(INFO - \[t1.1\] -(current:running)(polled) started\).*$/\1/p' \
     "${SUITE_RUN_DIR}/log/suite/log" >'sed-log.out'
 contains_ok 'sed-log.out' <<'__LOG__'
-[jobs-poll err] Connection refused
+[jobs-poll err]
+Connection refused
+__LOG__
+contains_ok 'sed-log.out' <<'__LOG__'
 INFO - [t1.1] -(current:running)(polled) started
 __LOG__
